scraper/pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = ""
authors = ["marko154 <marko.gartnar777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.5"
python-dotenv = "^1.0.1"
httpx = "^0.28.1"
xmltodict = "^0.14.2"
tenacity = "^9.0.0"
crawl4ai = "^0.5.0.post8"
feedparser = "^6.0.11"
langchain = "^0.3.23"
langchain-openai = "^0.3.12"
langchain-google-genai = "^2.1.2"
tiktoken = "^0.9.0"
html2text = "^2025.4.15"
python-readability = "^0.1.3"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
enable_incomplete_feature = "NewGenericSyntax"
